+++ src/main/java/com/github/argon4w/acceleratedrendering/core/mixins/LevelRendererMixin.java
@@ -79,6 +79,10 @@ public class LevelRendererMixin {
 			Matrix4f		pProjectionMatrix,
 			CallbackInfo	ci
 	) {
+		if (!CoreFeature.isLoaded()) {
+			return;
+		}
+
 		CoreStates							.recordBuffers	();
 		CoreBuffers.POS_TEX_COLOR_OUTLINE	.prepareBuffers	();
 		CoreStates							.restoreBuffers	();
@@ -95,6 +99,10 @@ public class LevelRendererMixin {
 			)
 	)
 	public void drawCoreBuffers(MultiBufferSource.BufferSource instance, Operation<Void> original) {
+		if (!CoreFeature.isLoaded()) {
+			return;
+		}
+
 		CoreStates						.recordBuffers	();
 		CoreBuffers.ENTITY				.prepareBuffers	();
 		CoreBuffers.BLOCK				.prepareBuffers	();
@@ -129,6 +137,10 @@ public class LevelRendererMixin {
 			at		= @At("TAIL")
 	)
 	public void deleteBuffers(CallbackInfo ci) {
+		if (!CoreFeature.isLoaded()) {
+			return;
+		}
+
 		CoreBuffers.ENTITY				.delete();
 		CoreBuffers.BLOCK				.delete();
 		CoreBuffers.POS					.delete();
